+++ spec/graphlient/adapters/http/faraday_adapter_spec.rb
@@ -54,7 +54,7 @@ describe Graphlient::Adapters::HTTP::FaradayAdapter do
     end
 
     it 'retrieves schema' do
-      expect(client.schema).to be_a GraphQL::Schema
+      expect(client.schema).to be_a Graphlient::Schema
     end
   end
 end