+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.3.1 (Next)
 
+* [#43](https://github.com/ashkan18/graphlient/pull/38): Allow to load and dump schema to json - [@povilasjurcys](https://github.com/povilasjurcys).
 * Your contribution here.
 
 ### 0.3.0 (02/22/2018)